requirements.txt
requests
google-search-results
openai
python-dotenv
flask
streamlit
pandas
schedule
